+++ app/build.gradle
@@ -70,6 +70,8 @@ dependencies {
     implementation "androidx.compose.material:material:1.3.1"
     implementation "androidx.compose.material3:material3:$material3_version"
 
+    implementation "com.google.accompanist:accompanist-systemuicontroller:0.28.0"
+
     def shizuku_version = "13.1.0"
     implementation "dev.rikka.shizuku:api:$shizuku_version"
     implementation "dev.rikka.shizuku:provider:$shizuku_version"
