+++ app/src/main/java/dev/bluehouse/enablevolte/pages/Config.kt
@@ -1,0 +1,76 @@
+package dev.bluehouse.enablevolte.pages
+
+import android.telephony.CarrierConfigManager
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.Dp
+import androidx.lifecycle.Lifecycle
+import androidx.navigation.NavController
+import dev.bluehouse.enablevolte.*
+import java.lang.IllegalStateException
+
+
+@Composable
+fun Config(navController: NavController) {
+    val modder = CarrierModder(LocalContext.current)
+
+    var configurable by rememberSaveable { mutableStateOf(false) }
+    var voLTEEnabled by rememberSaveable { mutableStateOf(false) }
+    var voWiFiEnabled by rememberSaveable { mutableStateOf(false) }
+    var configuredUserAgent by rememberSaveable { mutableStateOf("") }
+
+    OnLifecycleEvent { _, event ->
+        if (event == Lifecycle.Event.ON_CREATE) {
+            configurable = try {
+                if (checkShizukuPermission(0)) {
+                    val c = modder.deviceSupportsIMS && modder.subscriptionId >= 0
+                    if (c) {
+                        voLTEEnabled = modder.isVolteConfigEnabled
+                        voWiFiEnabled = modder.isVowifiConfigEnabled
+                        configuredUserAgent = modder.userAgentConfig
+                    }
+                    c
+                } else {
+                    false
+                }
+            } catch (e: IllegalStateException) {
+                false
+            }
+        }
+    }
+
+    Column(modifier = Modifier.padding(Dp(16f))) {
+        HeaderText(text = "Toggles")
+        BooleanPropertyView(label = "Enable VoLTE", toggled = voLTEEnabled) {
+            voLTEEnabled = if (voLTEEnabled) {
+                modder.updateCarrierConfig(CarrierConfigManager.KEY_CARRIER_VOLTE_AVAILABLE_BOOL, false)
+                false
+            } else {
+                modder.updateCarrierConfig(CarrierConfigManager.KEY_CARRIER_VOLTE_AVAILABLE_BOOL, true)
+                true
+            }
+        }
+        BooleanPropertyView(label = "Enable VoWiFi", toggled = voWiFiEnabled) {
+            voWiFiEnabled = if (voWiFiEnabled) {
+                modder.updateCarrierConfig(CarrierConfigManager.KEY_CARRIER_WFC_IMS_AVAILABLE_BOOL, false)
+                false
+            } else {
+                modder.updateCarrierConfig(CarrierConfigManager.KEY_CARRIER_WFC_IMS_AVAILABLE_BOOL, true)
+                true
+            }
+        }
+
+        HeaderText(text = "String Values")
+        StringPropertyView(label = "User Agent", value = configuredUserAgent) {
+            modder.updateCarrierConfig(modder.KEY_IMS_USER_AGENT, it)
+            configuredUserAgent = it
+        }
+    }
+}
